Refactor Hydro functions to their own manager
